Move `components.draw_sine_wave()` to `components.waves.sine()`
